+++ beaker/services/dataset.py
@@ -488,6 +488,17 @@ class DatasetClient(ServiceClient):
             ).json()
         )
 
+    def url(self, dataset: Union[str, Dataset]) -> str:
+        """
+        Get the URL for a dataset.
+
+        :param dataset: The dataset ID, name, or object.
+
+        :raises DatasetNotFound: If the dataset can't be found.
+        """
+        dataset_id = self.resolve_dataset(dataset).id
+        return f"{self.config.agent_address}/ds/{self.url_quote(dataset_id)}/details"
+
     def _not_found_err_msg(self, dataset: Union[str, Dataset]) -> str:
         dataset = dataset if isinstance(dataset, str) else dataset.id
         return (