+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+- Added `Beaker.dataset.url()` method.
+
 ### Changed
 
 - Improved performance of `Beaker.cluster.filter_available()` by using a `ThreadPoolExecutor` for concurrency.